refactor(catalog): factor out main command
Factor the content of main() into an exported function in a separate
package so that it can be invoked externally.
